backend/requirements.txt
# ── Core Framework ───────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── Database ────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# ── Authentication ──────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── Configuration ───────────────────────────────────────────────────────────
pydantic-settings>=2.1.0

# ── Redis ───────────────────────────────────────────────────────────────────
redis>=5.0.0

# ── RabbitMQ ────────────────────────────────────────────────────────────────
aio-pika>=9.4.0

# ── HTTP Client ─────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Request Parsing ─────────────────────────────────────────────────────────
python-multipart>=0.0.9

# ── Server-Sent Events ──────────────────────────────────────────────────────
sse-starlette>=2.0.0

# ── Task Queue ──────────────────────────────────────────────────────────────
celery>=5.3.0

# ── Tests ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
pytest-cov>=5.0.0
aiosqlite>=0.20.0
